+++ Sources/CodexBarCore/OpenAIWeb/ChromeCookieImporter.swift
@@ -1,3 +1,4 @@
+#if os(macOS)
 import CommonCrypto
 import Foundation
 import Security
@@ -426,3 +427,4 @@ enum ChromeCookieImporter {
         }
     }
 }
+#endif